refactoring v3
