+++ lib/navigation/routers.dart
@@ -1,18 +1,18 @@
 import 'package:get/get.dart';
 
-import '../ui/screens/barcode/home.dart';
-import '../ui/screens/easy_ship/home.dart';
-import '../ui/screens/enroll/enroll.dart';
-import '../ui/screens/enroll/screens/enroll_confirmation/home.dart';
-import '../ui/screens/enroll/screens/order_complete/enroll_complete.dart';
+import '../ui/screens/barcode/barcode.screen.dart';
+import '../ui/screens/easy_ship/easyship.screen.dart';
+import '../ui/screens/enroll/enrollhome.screen.dart';
+import '../ui/screens/enroll/screens/enroll_confirmation/enrollconfirmation.screen.dart';
+import '../ui/screens/enroll/screens/order_complete/enrollcomplete.screen.dart';
 import '../ui/screens/home/home.dart';
-import '../ui/screens/login/login.dart';
-import '../ui/screens/open_po/home/home.dart';
-import '../ui/screens/order_entry/order_entry.dart';
-import '../ui/screens/order_entry/screens/checkout/checkout_screen.dart';
-import '../ui/screens/sales_reports/home.dart';
-import '../ui/screens/splash/screen.dart';
-import '../ui/screens/webview/webview.dart';
+import '../ui/screens/login/login.screen.dart';
+import '../ui/screens/open_po/home/home.screen.dart';
+import '../ui/screens/order_entry/orderentry.screen.dart';
+import '../ui/screens/order_entry/screens/checkout/checkout.screen.dart';
+import '../ui/screens/sales_reports/salesreports.screen.dart';
+import '../ui/screens/splash/splash.screen.dart';
+import '../ui/screens/webview/webview.screen.dart';
 
 class AppRoutes {
   AppRoutes._(); //this is to prevent anyone from instantiating this object
