+++ lib/ui/screens/barcode/barcode.screen.dart
