+++ lib/ui/screens/barcode/controller/barcode.controller.dart
