+++ lib/ui/screens/enroll/screens/order_complete/enrollcomplete.screen.dart
@@ -4,7 +4,7 @@ import 'package:get/get.dart';
 import '../../../../../constants/colors.dart';
 import '../../../../../ui/global/widgets/transparent_app_bar.dart';
 import '../../../../../utilities/size_config.dart';
-import '../../bloc/enroll_complete_controller.dart';
+import '../../bloc/enrollcomplete.controller.dart';
 import 'components/body.dart';
 
 class EnrollComplete extends StatelessWidget {
