+++ lib/ui/screens/inventory/controller/inventory.controller.dart
