+++ lib/ui/screens/order_entry/controllers/orderentry.checkout.controller.dart
@@ -14,7 +14,7 @@ import 'package:code_magic_ex/models/radio_button_value.dart';
 import 'package:code_magic_ex/models/user_info.dart';
 import 'package:code_magic_ex/models/user_minimal_data.dart';
 import 'package:code_magic_ex/ui/global/widgets/overlay_progress.dart';
-import 'package:code_magic_ex/ui/screens/order_entry/screens/order_complete/order_complete.dart';
+import 'package:code_magic_ex/ui/screens/order_entry/screens/order_complete/ordercomplete.screen.dart';
 import 'package:code_magic_ex/utilities/logger.dart';
 import 'package:code_magic_ex/utilities/constants.dart';
 import 'package:code_magic_ex/utilities/function.dart';
