+++ lib/utilities/function.dart
@@ -19,7 +19,7 @@ import '../exceptions/unauthorised.exception.dart';
 import '../models/inventory_records.dart';
 import '../models/locale.dart';
 import '../ui/global/widgets/overlay_progress.dart';
-import '../ui/screens/login/login.dart';
+import '../ui/screens/login/login.screen.dart';
 import 'enums.dart';
 import 'logger.dart';
 import 'parsing.dart';
